refactor: modularize chatbot components and update integration
