+++ SortVision/src/App.jsx
@@ -6,8 +6,9 @@ import { getAlgorithmMetaTags, getHomepageMetaTags, getContributionsMetaTags, ge
 import SEOContent from './components/SEOContent';
 import { FeedbackButton } from './components/feedback';
 import { SettingsButton } from './components/settings';
-import ChatAssistant from './components/ChatAssistant';
+import { ChatAssistant } from "@/components/chatbot";
 import { AlgorithmStateProvider } from "./context/AlgorithmState";
+import { MobileOverlayContext } from '@/components/MobileOverlay';
 
 
 
@@ -383,209 +384,213 @@ const App = () => {
     </div>
   ), []);
 
+  const [isMobileOverlayVisible, setMobileOverlayVisible] = useState(false);
+
   return (
-    <AlgorithmStateProvider>
-      <div className="min-h-screen bg-slate-950 flex flex-col items-center justify-center p-2 sm:p-5 overflow-hidden">
-        {/* Mobile Detection Overlay - Lazy loaded */}
-        <Suspense fallback={null}>
-          <MobileOverlay />
-        </Suspense>
-
-        {/* SEO Helmet */}
-        <Helmet>
-          <title>{metaTags.title}</title>
-          <meta name="description" content={metaTags.description} />
-          <meta name="keywords" content={metaTags.keywords} />
-          <meta property="article:modified_time" content={currentDate} />
-
-          {/* Open Graph / Facebook */}
-          <meta property="og:type" content="website" />
-          <meta property="og:url" content={canonicalUrl} />
-          <meta property="og:title" content={metaTags.ogTitle} />
-          <meta property="og:description" content={metaTags.ogDescription} />
-          <meta property="og:updated_time" content={currentDate} />
-
-          {/* Twitter */}
-          <meta name="twitter:url" content={canonicalUrl} />
-          <meta name="twitter:title" content={metaTags.twitterTitle} />
-          <meta name="twitter:description" content={metaTags.twitterDescription} />
-
-          {/* Schema.org markup for Google */}
-          <script type="application/ld+json">
-            {JSON.stringify(schemaMarkup)}
-          </script>
-        </Helmet>
-
-        <SettingsButton />
-
-        {/* Header with logo and title */}
-        <Header>
-          <div className="flex items-center gap-2 sm:gap-3">
-            <Terminal className="h-6 w-6 sm:h-8 sm:w-8 text-emerald-400 animate-pulse animate-infinite animate-duration-[3000ms]" aria-hidden="true" />
-            <h1 className="text-2xl sm:text-4xl font-mono font-bold text-white">
-              <Link to="/" className="hover:opacity-90 transition-opacity">
-                <span className="text-emerald-400 hover:text-emerald-300 transition-colors duration-300">Sort</span>
-                <span className="text-purple-400 hover:text-purple-300 transition-colors duration-300">Vision</span>
-              </Link>
-            </h1>
-            <Code className="h-4 w-4 sm:h-6 sm:w-6 text-slate-400 animate-spin animate-once animate-duration-[1500ms] animate-delay-300" aria-hidden="true" />
-          </div>
-          <div className="text-lg sm:text-xl font-mono text-slate-400 mt-1">
-            <span className="text-emerald-400 hover:text-emerald-300 transition-colors duration-300">algorithm</span>
-            <span className="text-purple-400 hover:text-purple-300 transition-colors duration-300">.visualizer</span>
-            <span className="text-slate-400 hover:text-white transition-colors duration-300">()</span>
-          </div>
-        </Header>
-
-        {/* Subtitle with typing animation */}
-        <div className="text-center text-slate-400 font-mono mb-6 sm:mb-8 max-w-[90%] sm:max-w-md h-6 animate-fade-up animate-once animate-duration-[800ms] animate-delay-300">
-          <span className="text-amber-400">//</span> {displayText}
-          {!isTypingComplete && <span className="inline-block w-2 h-4 bg-amber-400 ml-1 animate-pulse" aria-hidden="true"></span>}
-        </div>
-
-        {/* Main Sorting Visualizer Component - Lazy loaded */}
-        <main className="animate-fade-up animate-once animate-duration-[1000ms] animate-delay-500 w-full max-w-4xl px-2 sm:px-4">
-          <h2 className="text-xl sm:text-2xl font-mono font-bold text-emerald-400 mb-4 text-center">
-            {algorithmName ? `${algorithmTitle} Visualization` : 'Sorting Algorithm Visualizer'}
-          </h2>
-          <Suspense fallback={fallbackElement}>
-            <SortingVisualizer
-              initialAlgorithm={currentAlgorithm}
-              activeTab={activeTab}
-              onTabChange={(newTab) => {
-                setActiveTab(newTab);
-
-                // Handle path-based routing for tab changes
-                if (specialMode === 'contributors') {
-                  // Handle contribution tab changes
-                  const sectionMapping = {
-                    'overview': 'overview',
-                    'guide': 'guide',
-                    'ssoc': 'ssoc'
-                  };
-                  const section = sectionMapping[newTab] || 'overview';
-                  navigate(`/contributions/${section}`, { replace: true });
-                } else {
-                  // Handle algorithm tab changes
-                  const pathMapping = {
-                    'controls': 'config',
-                    'metrics': 'metrics',
-                    'details': 'details'
-                  };
-                  const pathSegment = pathMapping[newTab] || 'config';
-                  const currentParams = new URLSearchParams(location.search);
-                  const newUrl = `/algorithms/${pathSegment}/${currentAlgorithm}${currentParams.toString() ? `?${currentParams.toString()}` : ''}`;
-                  navigate(newUrl, { replace: true });
-                }
-              }}
-              specialMode={specialMode}
-            />
+    <MobileOverlayContext.Provider value={{ isMobileOverlayVisible, setMobileOverlayVisible }}>
+      <AlgorithmStateProvider>
+        <div className="min-h-screen bg-slate-950 flex flex-col items-center justify-center p-2 sm:p-5 overflow-hidden">
+          {/* Mobile Detection Overlay - Lazy loaded */}
+          <Suspense fallback={null}>
+            <MobileOverlay />
           </Suspense>
-        </main>
 
-        {/* Footer */}
-        <Footer>
-          <span className="text-slate-600">/**</span> Built with
-          <span className="inline-block animate-bounce animate-infinite animate-duration-[2000ms] mx-1" aria-hidden="true">❤️</span>
-          by alienX <span className="text-slate-600">*/</span>
+          {/* SEO Helmet */}
+          <Helmet>
+            <title>{metaTags.title}</title>
+            <meta name="description" content={metaTags.description} />
+            <meta name="keywords" content={metaTags.keywords} />
+            <meta property="article:modified_time" content={currentDate} />
 
-          {/* Social links - Now wraps on mobile */}
-          <div className="mt-2 flex flex-wrap items-center justify-center gap-2 sm:gap-4 px-2 sm:px-4">
-            <button
-              onClick={() => {
-                if (specialMode === 'contributors') {
-                  // Return to normal mode - go to algorithms
-                  if (currentAlgorithm) {
-                    navigate(`/algorithms/config/${currentAlgorithm}`);
-                  } else {
-                    navigate('/algorithms/config/bubble'); // Default to bubble sort
-                  }
-                } else {
-                  // Go to contributors mode - navigate to contributions page
-                  navigate('/contributions/overview');
-                }
-              }}
-              className="flex items-center gap-1 text-slate-400 hover:text-indigo-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
-              aria-label={specialMode === 'contributors' ? "Return to SortVision main interface" : "View SortVision contributors"}
-            >
-              {specialMode === 'contributors' ? (
-                <Terminal className="h-3 w-3 sm:h-4 sm:w-4" aria-hidden="true" />
-              ) : (
-                <Users className="h-3 w-3 sm:h-4 sm:w-4" aria-hidden="true" />
-              )}
-              <span>{specialMode === 'contributors' ? 'SortVision' : 'Contributors'}</span>
-            </button>
+            {/* Open Graph / Facebook */}
+            <meta property="og:type" content="website" />
+            <meta property="og:url" content={canonicalUrl} />
+            <meta property="og:title" content={metaTags.ogTitle} />
+            <meta property="og:description" content={metaTags.ogDescription} />
+            <meta property="og:updated_time" content={currentDate} />
 
-            <a
-              href="https://github.com/alienx5499/SortVision"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="flex items-center gap-1 text-slate-400 hover:text-emerald-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
-              aria-label="View SortVision source code on GitHub"
-            >
-              <Github className="h-3 w-3 sm:h-4 sm:w-4" aria-hidden="true" />
-              <span>GitHub</span>
-            </a>
+            {/* Twitter */}
+            <meta name="twitter:url" content={canonicalUrl} />
+            <meta name="twitter:title" content={metaTags.twitterTitle} />
+            <meta name="twitter:description" content={metaTags.twitterDescription} />
 
-            <a
-              href="https://www.linkedin.com/in/prabalpatra5499/"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="flex items-center gap-1 text-slate-400 hover:text-blue-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
-              aria-label="Connect with the developer on LinkedIn"
-            >
-              <Linkedin className="h-3 w-3 sm:h-4 sm:w-4" aria-hidden="true" />
-              <span>LinkedIn</span>
-            </a>
+            {/* Schema.org markup for Google */}
+            <script type="application/ld+json">
+              {JSON.stringify(schemaMarkup)}
+            </script>
+          </Helmet>
 
-            <a
-              href="https://github.com/sponsors/alienx5499"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="flex items-center gap-1 text-slate-400 hover:text-pink-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
-            >
-              <span className="text-base sm:text-lg">♥</span>
-              <span>Sponsor</span>
-            </a>
+          <SettingsButton />
 
-            <a
-              href="https://buymeacoffee.com/alienx5499"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="flex items-center gap-1 text-slate-400 hover:text-yellow-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
-              aria-label="Support the developer with a donation"
-            >
-              <span className="text-base sm:text-lg" aria-hidden="true">☕</span>
-              <span>Buy me a coffee</span>
-            </a>
+          {/* Header with logo and title */}
+          <Header>
+            <div className="flex items-center gap-2 sm:gap-3">
+              <Terminal className="h-6 w-6 sm:h-8 sm:w-8 text-emerald-400 animate-pulse animate-infinite animate-duration-[3000ms]" aria-hidden="true" />
+              <h1 className="text-2xl sm:text-4xl font-mono font-bold text-white">
+                <Link to="/" className="hover:opacity-90 transition-opacity">
+                  <span className="text-emerald-400 hover:text-emerald-300 transition-colors duration-300">Sort</span>
+                  <span className="text-purple-400 hover:text-purple-300 transition-colors duration-300">Vision</span>
+                </Link>
+              </h1>
+              <Code className="h-4 w-4 sm:h-6 sm:w-6 text-slate-400 animate-spin animate-once animate-duration-[1500ms] animate-delay-300" aria-hidden="true" />
+            </div>
+            <div className="text-lg sm:text-xl font-mono text-slate-400 mt-1">
+              <span className="text-emerald-400 hover:text-emerald-300 transition-colors duration-300">algorithm</span>
+              <span className="text-purple-400 hover:text-purple-300 transition-colors duration-300">.visualizer</span>
+              <span className="text-slate-400 hover:text-white transition-colors duration-300">()</span>
+            </div>
+          </Header>
 
-            <a
-              href="https://x.com/alienx5499"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="flex items-center gap-1 text-slate-400 hover:text-sky-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
-              aria-label="Follow the developer on X (Twitter)"
-            >
-              <Twitter className="h-3 w-3 sm:h-4 sm:w-4" aria-hidden="true" />
-              <span>Twitter</span>
-            </a>
+          {/* Subtitle with typing animation */}
+          <div className="text-center text-slate-400 font-mono mb-6 sm:mb-8 max-w-[90%] sm:max-w-md h-6 animate-fade-up animate-once animate-duration-[800ms] animate-delay-300">
+            <span className="text-amber-400">//</span> {displayText}
+            {!isTypingComplete && <span className="inline-block w-2 h-4 bg-amber-400 ml-1 animate-pulse" aria-hidden="true"></span>}
           </div>
 
-          {/* Assistant Chatbot */}
-          <ChatAssistant />
+          {/* Main Sorting Visualizer Component - Lazy loaded */}
+          <main className="animate-fade-up animate-once animate-duration-[1000ms] animate-delay-500 w-full max-w-4xl px-2 sm:px-4">
+            <h2 className="text-xl sm:text-2xl font-mono font-bold text-emerald-400 mb-4 text-center">
+              {algorithmName ? `${algorithmTitle} Visualization` : 'Sorting Algorithm Visualizer'}
+            </h2>
+            <Suspense fallback={fallbackElement}>
+              <SortingVisualizer
+                initialAlgorithm={currentAlgorithm}
+                activeTab={activeTab}
+                onTabChange={(newTab) => {
+                  setActiveTab(newTab);
 
-          {/* Other Components */}
+                  // Handle path-based routing for tab changes
+                  if (specialMode === 'contributors') {
+                    // Handle contribution tab changes
+                    const sectionMapping = {
+                      'overview': 'overview',
+                      'guide': 'guide',
+                      'ssoc': 'ssoc'
+                    };
+                    const section = sectionMapping[newTab] || 'overview';
+                    navigate(`/contributions/${section}`, { replace: true });
+                  } else {
+                    // Handle algorithm tab changes
+                    const pathMapping = {
+                      'controls': 'config',
+                      'metrics': 'metrics',
+                      'details': 'details'
+                    };
+                    const pathSegment = pathMapping[newTab] || 'config';
+                    const currentParams = new URLSearchParams(location.search);
+                    const newUrl = `/algorithms/${pathSegment}/${currentAlgorithm}${currentParams.toString() ? `?${currentParams.toString()}` : ''}`;
+                    navigate(newUrl, { replace: true });
+                  }
+                }}
+                specialMode={specialMode}
+              />
+            </Suspense>
+          </main>
+
+          {/* Footer */}
+          <Footer>
+            <span className="text-slate-600">/**</span> Built with
+            <span className="inline-block animate-bounce animate-infinite animate-duration-[2000ms] mx-1" aria-hidden="true">❤️</span>
+            by alienX <span className="text-slate-600">*/</span>
+
+            {/* Social links - Now wraps on mobile */}
+            <div className="mt-2 flex flex-wrap items-center justify-center gap-2 sm:gap-4 px-2 sm:px-4">
+              <button
+                onClick={() => {
+                  if (specialMode === 'contributors') {
+                    // Return to normal mode - go to algorithms
+                    if (currentAlgorithm) {
+                      navigate(`/algorithms/config/${currentAlgorithm}`);
+                    } else {
+                      navigate('/algorithms/config/bubble'); // Default to bubble sort
+                    }
+                  } else {
+                    // Go to contributors mode - navigate to contributions page
+                    navigate('/contributions/overview');
+                  }
+                }}
+                className="flex items-center gap-1 text-slate-400 hover:text-indigo-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
+                aria-label={specialMode === 'contributors' ? "Return to SortVision main interface" : "View SortVision contributors"}
+              >
+                {specialMode === 'contributors' ? (
+                  <Terminal className="h-3 w-3 sm:h-4 sm:w-4" aria-hidden="true" />
+                ) : (
+                  <Users className="h-3 w-3 sm:h-4 sm:w-4" aria-hidden="true" />
+                )}
+                <span>{specialMode === 'contributors' ? 'SortVision' : 'Contributors'}</span>
+              </button>
+
+              <a
+                href="https://github.com/alienx5499/SortVision"
+                target="_blank"
+                rel="noopener noreferrer"
+                className="flex items-center gap-1 text-slate-400 hover:text-emerald-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
+                aria-label="View SortVision source code on GitHub"
+              >
+                <Github className="h-3 w-3 sm:h-4 sm:w-4" aria-hidden="true" />
+                <span>GitHub</span>
+              </a>
+
+              <a
+                href="https://www.linkedin.com/in/prabalpatra5499/"
+                target="_blank"
+                rel="noopener noreferrer"
+                className="flex items-center gap-1 text-slate-400 hover:text-blue-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
+                aria-label="Connect with the developer on LinkedIn"
+              >
+                <Linkedin className="h-3 w-3 sm:h-4 sm:w-4" aria-hidden="true" />
+                <span>LinkedIn</span>
+              </a>
+
+              <a
+                href="https://github.com/sponsors/alienx5499"
+                target="_blank"
+                rel="noopener noreferrer"
+                className="flex items-center gap-1 text-slate-400 hover:text-pink-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
+              >
+                <span className="text-base sm:text-lg">♥</span>
+                <span>Sponsor</span>
+              </a>
+
+              <a
+                href="https://buymeacoffee.com/alienx5499"
+                target="_blank"
+                rel="noopener noreferrer"
+                className="flex items-center gap-1 text-slate-400 hover:text-yellow-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
+                aria-label="Support the developer with a donation"
+              >
+                <span className="text-base sm:text-lg" aria-hidden="true">☕</span>
+                <span>Buy me a coffee</span>
+              </a>
+
+              <a
+                href="https://x.com/alienx5499"
+                target="_blank"
+                rel="noopener noreferrer"
+                className="flex items-center gap-1 text-slate-400 hover:text-sky-400 hover:scale-110 transition-all duration-300 text-[10px] sm:text-xs"
+                aria-label="Follow the developer on X (Twitter)"
+              >
+                <Twitter className="h-3 w-3 sm:h-4 sm:w-4" aria-hidden="true" />
+                <span>Twitter</span>
+              </a>
+            </div>
+
+            {/* Assistant Chatbot */}
+            <ChatAssistant />
+
+            {/* Other Components */}
 
 
-        </Footer>
+          </Footer>
 
-        {/* SEO Content for better search engine understanding */}
-        <SEOContent algorithm={algorithmName} />
+          {/* SEO Content for better search engine understanding */}
+          <SEOContent algorithm={algorithmName} />
 
-        {/* Floating Feedback Button */}
-        <FeedbackButton />
-      </div>
-    </AlgorithmStateProvider>
+          {/* Floating Feedback Button */}
+          <FeedbackButton />
+        </div>
+      </AlgorithmStateProvider>
+    </MobileOverlayContext.Provider>
   );
 };
