+++ SortVision/src/components/chatbot/index.js
@@ -1,0 +1,4 @@
+export { default as ChatAssistant } from './ChatAssistant';
+export { default as ChatButton } from './ChatButton';
+export { default as ChatModal } from './ChatModal';
+export { AssistantEngine } from './assistantEngine';
